Data-transformation-with-spark/requirements.txt
mysql-connector-python==9.0.0
psycopg2-binary==2.9.9
pandas
pyspark==3.5.2